// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.31.2
 * To regenerate, run `npx convex dev`
 */

import type { FunctionReference } from "convex/server";

// Simplified type declarations to avoid TS2589 deep type instantiation errors

export declare const api: {
  posts: {
    list: FunctionReference<"query", "public">;
    get: FunctionReference<"query", "public">;
    getScheduled: FunctionReference<"query", "public">;
    getDrafts: FunctionReference<"query", "public">;
    create: FunctionReference<"mutation", "public">;
    update: FunctionReference<"mutation", "public">;
    remove: FunctionReference<"mutation", "public">;
    duplicate: FunctionReference<"mutation", "public">;
    schedule: FunctionReference<"mutation", "public">;
    publish: FunctionReference<"mutation", "public">;
    getStats: FunctionReference<"query", "public">;
  };
  analytics: {
    getDashboardStats: FunctionReference<"query", "public">;
    getEngagementTrends: FunctionReference<"query", "public">;
    getTopPosts: FunctionReference<"query", "public">;
    getFollowerGrowth: FunctionReference<"query", "public">;
    getPlatformStats: FunctionReference<"query", "public">;
    getUpcomingPosts: FunctionReference<"query", "public">;
    getRecommendations: FunctionReference<"query", "public">;
  };
};

export declare const internal: Record<string, unknown>;